requirements.txt
python-dotenv
certifi

pymongo
pymongo[srv] == 3.12

numpy
pandas

matplotlib
seaborn

scikit-learn
catboost
xgboost


flask
dill
pyaml

-e .